+++ example/lib/pages/notification_box_page.dart
@@ -1,5 +1,5 @@
-import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:flutter/material.dart';
+import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 
 import '../native_app.dart';
 
@@ -23,7 +23,7 @@ class _NotificationBoxPage extends State<NotificationBoxPage> {
         const ThemeModeSegmentedButton(),
         const SizedBox(height: sbbDefaultSpacing),
         const SBBListHeader('Alert'),
-        SBBGroup(
+        SBBContentBox(
           padding: const EdgeInsets.all(sbbDefaultSpacing),
           child: Column(
             children: [
@@ -52,7 +52,7 @@ class _NotificationBoxPage extends State<NotificationBoxPage> {
           ),
         ),
         const SBBListHeader('Warning'),
-        SBBGroup(
+        SBBContentBox(
           padding: const EdgeInsets.all(sbbDefaultSpacing),
           child: Column(
             children: [
@@ -80,7 +80,7 @@ class _NotificationBoxPage extends State<NotificationBoxPage> {
           ),
         ),
         const SBBListHeader('Success'),
-        SBBGroup(
+        SBBContentBox(
           padding: const EdgeInsets.all(sbbDefaultSpacing),
           child: Column(
             children: [
@@ -108,7 +108,7 @@ class _NotificationBoxPage extends State<NotificationBoxPage> {
           ),
         ),
         const SBBListHeader('Information'),
-        SBBGroup(
+        SBBContentBox(
           padding: const EdgeInsets.all(sbbDefaultSpacing),
           child: Column(
             children: [