+++ lib/sbb_design_system_mobile.dart
@@ -2,7 +2,7 @@ export 'src/accordion/sbb_accordion.dart';
 export 'src/button/button.dart';
 export 'src/checkbox/checkbox.dart';
 export 'src/chip/sbb_chip.dart';
-export 'src/group/sbb_group.dart';
+export 'src/container/sbb_content_box.dart';
 export 'src/header/sbb_header.dart';
 export 'src/headerbox/headerbox.dart';
 export 'src/input_trigger/sbb_input_trigger.dart';